requirements.txt
pyyaml
h5py
Pillow==7.2.0
flask
flask_restful
tensorflow_hub